Function amazon_sp_uploads::apis::uploads_api::create_upload_destination_for_resource

pub async fn create_upload_destination_for_resource(
    configuration: &Configuration,
    marketplace_ids: Vec<String>,
    content_md5: &str,
    resource: &str,
    content_type: Option<&str>
) -> Result<CreateUploadDestinationResponse, Error>

Creates an upload destination, returning the information required to upload a file to the destination and to programmatically access the file. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | .1 | 5 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.